Second composite example - added a collection of painters in the form of a painting company.

+++ src/main/java/composite/painter/Painter.java
@@ -10,7 +10,7 @@ public class Painter implements PainterInterface {
         this.daysPerHouse = daysPerHouse;
     }
 
-    public int estimateDaysToPaint(int numberOfHouses) {
+    public float estimateDaysToPaint(int numberOfHouses) {
         return numberOfHouses * daysPerHouse;
     }
 }
