+++ src/main/java/composite/painter/PainterInterface.java
@@ -4,5 +4,5 @@ package main.java.composite.painter;
  * Created by mpalanisamy on 8/17/16.
  */
 public interface PainterInterface {
-    public int estimateDaysToPaint(int numberOfHouses);
+    public float estimateDaysToPaint(int numberOfHouses);
 }
